+++ js/util.js
@@ -1,0 +1,19 @@
+const getRandomInt = (min, max) => Math.floor(Math.random() * (max - min + 1) + min);
+
+const getRandomIdFromRange = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInt(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInt(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+export {getRandomInt, getRandomIdFromRange};
